RIN-624 - Moves web services functionality into it's own module.
