+++ lib/connect-service/endpoint.js
