+++ lib/connect-service/get-auth-token.js
